requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
dj-email-url==1.0.6
Django==4.1.7
django-cache-url==3.4.4
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
environs==9.5.0
marshmallow==3.19.0
packaging==23.0
phonenumbers==8.13.8
python-dotenv==1.0.0
sqlparse==0.4.3
typing_extensions==4.5.0
